[kernel] Various source code cleanups
