+++ elks/arch/i86/kernel/system.c
@@ -61,14 +61,6 @@ void INITPROC setup_arch(seg_t *start, seg_t *end)
 
 void hard_reset_now(void)
 {
-#ifdef __BCC__
-    asm(\
-	"\tmov ax,#0x40\n" \
-	"\tmov ds, ax\n" \
-	"\tmov [0x72],#0x1234\n" \
-	"\tjmp #0xffff:0\n" \
-	);
-#endif
 #ifdef __ia16__
     asm("mov $0x40,%ax\n\t"
 	"mov %ax,%ds\n\t"
@@ -86,38 +78,20 @@ void hard_reset_now(void)
  */
 void apm_shutdown_now(void)
 {
-#ifdef __BCC__
-    asm(\
-        "\tmov ax,#0x5301\n" \
-        "\txor bx,bx\n" \
-        "\tint #0x15\n" \
-        "\tjc apm_error\n" \
-        "\tmov ax,#0x5308\n" \
-        "\tmov bx,#0x0001\n" \
-        "\tmov cx,#0x0001\n" \
-        "\tint #0x15\n" \
-        "\tjc apm_error\n" \
-        "\tmov ax,#0x5307\n" \
-        "\tmov bx,#0x0001\n" \
-        "\tmov cx,#0x0003\n" \
-        "\tint #0x15\n" \
-        "\tapm_error:\n"
-        );
-#endif
 #ifdef __ia16__
-    asm("movw $21249,%ax\n\t"
+    asm("movw $0x5301,%ax\n\t"
 	"xorw %bx,%bx\n\t"
-	"int $21\n\t"
+	"int $0x15\n\t"
 	"jc apm_error\n\t"
-	"movw $21256,%ax\n\t"
+	"movw $0x5308,%ax\n\t"
 	"movw $1,%bx\n\t"
 	"movw $1,%cx\n\t"
-	"int $21\n\t"
+	"int $0x15\n\t"
 	"jc apm_error\n\t"
-	"movw $21255,%ax\n\t"
+	"movw $0x5307,%ax\n\t"
 	"movw $1,%bx\n\t"
 	"movw $3,%cx\n\t"
-	"int $21\n\t"
+	"int $0x15\n\t"
 	"apm_error:\n\t"
 	);
 #endif
