+++ elks/arch/i86/lib/unused/memdumpk.c
