+++ elks/include/arch/asm.h
@@ -1,10 +1,6 @@
 #ifndef __ARCH_8086_ASM_H
 #define __ARCH_8086_ASM_H
 
-#ifdef S_SPLINT_S
-
 void asm(char *);
-
-#endif
 
 #endif
