+++ elks/include/arch/keyboard.h
@@ -5,10 +5,4 @@ struct pt_regs;
 extern void keyboard_irq(int,struct pt_regs *,void *);
 extern void Console_set_vc(unsigned);
 
-#ifdef CONFIG_ARCH_SIBO
-
-extern int psiongetchar(void);
-
-#endif
-
 #endif
