+++ elks/include/arch/param.h
@@ -3,14 +3,15 @@
 
 #include <linuxmt/config.h>
 
-#ifndef HZ
-
+/* We don't support the complex PLL time loops on Linux 8086
+ * as 100Hz fixed is more than good enough for us.
+ *
+ * set to 100 HZ by Shani <kerr@wizard.net>, since that's what we're using!
+ */
 #ifdef CONFIG_ARCH_SIBO
 #define HZ			30
 #else
 #define HZ			100
-#endif
-
 #endif
 
 #define RUNNABLE_PAGESIZE	4096
